backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2

# NLP and text processing
spacy==3.7.2
nltk==3.8.1
emoji==2.8.0

python-dotenv==1.0.0
logging
supabase==2.3.3